# openapi.yaml
openapi: "3.1.0"
info:
  title: WSI Viewer Backend
  version: "1.0.0"
  description: |
    Whole-slide image viewer — tile streaming, session management,
    Tumour Content Analysis (TCA), per-slide annotations, and pathology review worklist.
servers:
  - url: http://localhost:8080
    description: Local dev
...
components:
...
  schemas:
...
    WorklistItem:
      type: object
      properties:
        id:                 { type: string }
        patient_id:         { type: string }
        event_id:           { type: string }
        selected_slide_id:  { type: string }
        status:             { type: string, enum: [PENDING, IN_REVIEW, COMPLETE, FLAGGED, CANCELLED] }
        priority:           { type: integer, minimum: 1, maximum: 5 }
        assigned_to:        { type: string, nullable: true }
        notes:              { type: string, nullable: true }
        created_at:         { type: string, format: date-time }
        updated_at:         { type: string, format: date-time }
        viewer_params:
          type: object
          properties:
            patient_id:        { type: string }
            event_id:          { type: string }
            selected_slide_id: { type: string }
    Worklist:
      type: object
      properties:
        items:  { type: array, items: { $ref: '#/components/schemas/WorklistItem' } }
        total:  { type: integer }
        limit:  { type: integer }
        offset: { type: integer }
...
paths:
...
  /api/worklist:
    get:
      tags: [worklist]
      summary: List worklist items (paginated, filterable)
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [PENDING, IN_REVIEW, COMPLETE, FLAGGED, CANCELLED] }
        - name: assigned_to
          in: query
          schema: { type: string }
        - name: priority
          in: query
          schema: { type: integer, minimum: 1, maximum: 5 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Worklist' } } } }
    post:
      tags: [worklist]
      summary: Add case to worklist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patient_id, event_id, selected_slide_id]
              properties:
                patient_id:        { type: string }
                event_id:          { type: string }
                selected_slide_id: { type: string }
                priority:          { type: integer, minimum: 1, maximum: 5, default: 3 }
                assigned_to:       { type: string, nullable: true }
                notes:             { type: string, nullable: true }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/WorklistItem' } } } }
...
  /api/worklist/{item_id}:
    get:
      tags: [worklist]
      summary: Get single worklist item
      parameters:
        - name: item_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/WorklistItem' } } } }
        "404": { description: Not found }
    patch:
      tags: [worklist]
      summary: Update worklist item
      parameters:
        - name: item_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:      { type: string, enum: [PENDING, IN_REVIEW, COMPLETE, FLAGGED, CANCELLED] }
                priority:    { type: integer, minimum: 1, maximum: 5 }
                assigned_to: { type: string, nullable: true }
                notes:       { type: string, nullable: true }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/WorklistItem' } } } }
        "404": { description: Not found }
        "422": { description: Validation error }
    delete:
      tags: [worklist]
      summary: Remove worklist item
      parameters:
        - name: item_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
        "404": { description: Not found }